+++ src/main/java/com/calibermc/buildify/datagen/properties/GlassWindowTypes.java
@@ -1,0 +1,15 @@
+package com.calibermc.buildify.datagen.properties;
+
+public enum GlassWindowTypes {
+    ;
+
+    private final String name;
+
+    GlassWindowTypes(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}