+++ src/main/java/com/calibermc/buildify/util/compat/ModCompats.java
@@ -16,10 +16,12 @@ public class ModCompats {
     public static final boolean BYG;
     public static final boolean CALIBER;
     public static final boolean CHIPPED;
+    public static final boolean CREATE;
     public static final boolean DOUBLE_DOORS;
     public static final boolean MANYIDEAS_DOORS;
     public static final boolean NATURES_CANVAS;
     public static final boolean QUARK;
+    public static final boolean STONE_WORKS;
     public static final boolean REGIONS_UNEXPLORED;
 
     static {
@@ -27,10 +29,12 @@ public class ModCompats {
         BYG = ModList.get().isLoaded("byg");
         CALIBER = ModList.get().isLoaded("caliber");
         CHIPPED = ModList.get().isLoaded("chipped");
+        CREATE = ModList.get().isLoaded("create");
         DOUBLE_DOORS = ModList.get().isLoaded("doubledoors");
         MANYIDEAS_DOORS = ModList.get().isLoaded("manyideas_doors");
         NATURES_CANVAS = ModList.get().isLoaded("naturescanvas");
         QUARK = ModList.get().isLoaded("quark");
+        STONE_WORKS = ModList.get().isLoaded("stone_works");
         REGIONS_UNEXPLORED = ModList.get().isLoaded("regions_unexplored");
 
 
@@ -47,6 +51,9 @@ public class ModCompats {
         if (CHIPPED) {
             compatMODID.add("chipped");
         }
+        if (CREATE) {
+            compatMODID.add("create");
+        }
         if (DOUBLE_DOORS) {
             compatMODID.add("doubledoors");
         }
@@ -58,6 +65,9 @@ public class ModCompats {
         }
         if (QUARK) {
             compatMODID.add("quark");
+        }
+        if (STONE_WORKS) {
+            compatMODID.add("stone_works");
         }
         if (REGIONS_UNEXPLORED) {
             compatMODID.add("regions_unexplored");